+++ src/main.rs
@@ -13,7 +13,7 @@ use zkutil::circom_circuit::{
     proof_to_json_file, prove as prove2, proving_key_json_file, r1cs_from_bin_file, r1cs_from_json_file, verification_key_json_file,
     verify as verify2, witness_from_json_file, CircomCircuit, R1CS,
 };
-use zkutil::proof_system::ProofSystem;
+use zkutil::proofsys_type::ProofSystem;
 
 /// A tool to work with SNARK circuits generated by circom
 #[derive(Clap)]
